requirements.txt
streamlit==1.50.0
tensorflow==2.16.1
numpy<2
pillow==11.0.0
opencv-python==4.10.0.84
matplotlib==3.8.0
numpy<2
pandas==2.3.3
protobuf<4
